Fix : merge 버그 수정

+++ app/src/main/java/com/gst/gusto/review_write/ReviewDetailFragment.kt
@@ -98,7 +98,7 @@ class ReviewDetailFragment : Fragment() {
         var reviewId = 1
         gustoViewModel.myReviewId = reviewId.toLong()
 
-        gustoViewModel.getReview(reviewId){
+        gustoViewModel.getReview(reviewId.toLong()){
             result ->
             when(result){
                 0 -> {
